+++ src/constants.js
@@ -1,8 +1,0 @@
-module.exports = {
-	/**
-	 * Replace with your instance wide constants
-	 */
-	initConstants() {
-		this.defineConst('API_URL_BASE', 'https://api.stagetimer.io/v0')
-	},
-}